Refactoring on App Main names & pom.xml s up
