+++ account-service/src/main/java/pl/piomin/microservices/account/AccountServiceApplication.java
@@ -1,0 +1,15 @@
+package pl.piomin.microservices.account;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class AccountServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AccountServiceApplication.class, args);
+	}
+	
+}
