+++ customer-service/pom.xml
@@ -38,7 +38,7 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 				<configuration>
-					<mainClass>pl.piomin.microservices.customer.Application</mainClass>
+					<mainClass>pl.piomin.microservices.customer.CustomerServiceApplication</mainClass>
 					<addResources>true</addResources>
 				</configuration>
 				<executions>
