+++ customer-service/src/main/java/pl/piomin/microservices/customer/CustomerServiceApplication.java
@@ -1,0 +1,17 @@
+package pl.piomin.microservices.customer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.netflix.feign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+@EnableFeignClients
+public class CustomerServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CustomerServiceApplication.class, args);
+	}
+
+}
