+++ discovery-service/src/main/java/pl/piomin/microservices/eureka/DiscoveryServiceApplication.java
@@ -1,0 +1,15 @@
+package pl.piomin.microservices.eureka;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class DiscoveryServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DiscoveryServiceApplication.class, args);
+	}
+
+}
