+++ gateway-service/pom.xml
@@ -46,7 +46,7 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 				<configuration>
-					<mainClass>pl.piomin.microservices.edge.Application</mainClass>
+					<mainClass>pl.piomin.microservices.edge.GatewayServiceApplication</mainClass>
 					<addResources>true</addResources>
 				</configuration>
 				<executions>
