--- gateway-service/src/main/java/pl/piomin/microservices/edge/Application.java
@@ -1,13 +1,0 @@
-package pl.piomin.microservices.edge;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-
-@SpringBootApplication
-@EnableZuulProxy
-public class Application {
-    public static void main(String[] args) {
-        new SpringApplicationBuilder(Application.class).web(true).run(args);
-    }
-}
