--- zipkin-service/src/main/java/pl/piomin/microservices/zipkin/Application.java
@@ -1,16 +1,0 @@
-package pl.piomin.microservices.zipkin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import zipkin.server.EnableZipkinServer;
-
-@SpringBootApplication
-@EnableZipkinServer
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-
-}
