+++ zipkin-service/src/main/java/pl/piomin/microservices/zipkin/ZipkinServiceApplication.java
@@ -1,0 +1,16 @@
+package pl.piomin.microservices.zipkin;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import zipkin.server.EnableZipkinServer;
+
+@SpringBootApplication
+@EnableZipkinServer
+public class ZipkinServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ZipkinServiceApplication.class, args);
+	}
+
+}
